Homework 3 was done

+++ Products.java
@@ -1,13 +1,13 @@
 public abstract class Products {
     protected String name;
-    protected int volume;
+    protected long volume;
 
     public abstract String getName();
 
     public abstract void setName(String name);
 
-    public abstract int getVolume();
+    public abstract Long getVolume();
 
-    public abstract void setVolume(int volume);
+    public abstract void setVolume(Long volume);
 
 }
